requirements.txt
# Core dependencies
pandas>=1.5.0
numpy>=1.21.0
yfinance>=0.2.0
requests>=2.28.0
python-dotenv>=0.19.0
pyyaml>=6.0.0

# Web Framework
flask==3.0.0
flask-socketio==5.3.6
python-socketio==5.11.1
python-engineio==4.9.0

# Trading API
kiteconnect>=4.0.0

# Database - Supabase
supabase==2.18.1
pyjwt==2.10.1

# WebSocket support for TrueData
websocket-client>=1.4.0
websockets>=11.0.0
aiohttp>=3.8.0

# Database
sqlalchemy>=1.4.0
psycopg2-binary>=2.9.0
# sqlite3 is built-in with Python

# Security and Authentication
cryptography>=41.0.0
bcrypt>=4.0.0

# Machine Learning
scikit-learn>=1.1.0
tensorflow>=2.10.0
torch>=1.12.0
torchvision>=0.13.0
keras>=2.10.0
xgboost>=1.6.0
lightgbm>=3.3.0
catboost>=1.1.0
prophet>=1.1.0
statsmodels>=0.13.0

# Deep Learning for Time Series
pytorch-lightning>=1.7.0
transformers>=4.21.0
optuna>=3.0.0

# Feature Engineering & Technical Analysis
ta-lib>=0.4.24
TA-Lib>=0.4.24
finta>=1.3
alpha-vantage>=2.3.0
pandas-ta>=0.3.14b0

# Advanced Analytics
scipy>=1.11.1
numpy>=1.24.0
pandas>=2.0.3
matplotlib>=3.5.0
seaborn>=0.11.0
plotly>=5.10.0

# Time Series Analysis
arch>=5.3.0
pmdarima>=2.0.0
sktime>=0.13.0
tsfresh>=0.19.0

# Model Optimization
hyperopt>=0.2.7
bayesian-optimization>=1.4.0
scikit-optimize>=0.9.0

# Model Persistence
joblib>=1.2.0
pickle5>=0.0.12
dill>=0.3.6

# GPU Support (Optional)
cupy-cuda11x>=11.0.0  # For NVIDIA CUDA support
numba>=0.56.0

# Visualization
matplotlib>=3.5.0
plotly>=5.10.0
seaborn>=0.11.0
streamlit>=1.15.0

# Trading APIs
kiteconnect>=4.0.0
upstox-python-sdk>=1.0.0
pyotp>=2.6.0
smartapi-python>=1.0.0

# Notifications
twilio>=7.16.0
schedule>=1.1.0

# Logging and Monitoring
structlog>=23.0.0
colorama>=0.4.0
psutil>=5.9.0

# Time and Date handling
pytz>=2023.0
APScheduler>=3.10.0

# Data Validation
marshmallow>=3.20.0
jsonschema>=4.19.0

# Testing
pytest>=7.0.0
pytest-cov>=4.0.0

# Deployment
docker>=6.0.0
gunicorn>=20.1.0

# Development
black>=22.0.0
flake8>=5.0.0
pre-commit>=2.20.0
